move CacheType enum into its own file
