+++ lib/streamlit/runtime/caching/cache_errors.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import enum
 import types
 from typing import Any, Optional
 
@@ -22,22 +21,7 @@ from streamlit.errors import (
     StreamlitAPIException,
     StreamlitAPIWarning,
 )
-
-
-class CacheType(enum.Enum):
-    """The function cache types we implement."""
-
-    DATA = "DATA"
-    RESOURCE = "RESOURCE"
-
-
-def get_decorator_api_name(cache_type: CacheType) -> str:
-    """Return the name of the public decorator API for the given CacheType."""
-    if cache_type is CacheType.DATA:
-        return "cache_data"
-    if cache_type is CacheType.RESOURCE:
-        return "cache_resource"
-    raise RuntimeError(f"Unrecognized CacheType '{cache_type}'")
+from streamlit.runtime.caching.cache_type import CacheType, get_decorator_api_name
 
 
 def get_cached_func_name_md(func: Any) -> str:
